requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
aiofiles